move ClusterInfo type to ANTISMASH module
